requirements.txt
Flask
rembg
Pillow
cloudinary